Add files via upload
This macro is still a work in progress.  I requires CE9.6 and above so that you can disable the share button.  The problem is that once you do that, you cannot stop sharing, so my next step when I have time for this macro will be to build in a stop sharing button.  The only downside is that you cannot have the stop sharing macro only appear when you are already actively sharing.