Update pinToCombine.js
When I was matching my PIN to the text response, the macro would repeat itself the number of times it has happened.  So after the second attempt, it would repeat itself twice.  After the third attempt, it would repeat itself three times.

To stop this, I set up the pin matching within it's own listener that waits for the textinput response.

I could put the showPinPad function within the panel clicked listener, but it's not a bad idea to have it here.  If I want to change the script to reopen the pin pad if they get it wrong, I could call the function back.